Rename filetypes
